requirements.txt
flask==2.0.3
flask-sqlalchemy==2.5.1
gunicorn==20.1.0
psycopg2==2.9.3
python-dotenv==0.20.0
setuptools==62.1.0
